requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-rest-framework==3.14.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
